// packages/bedrock-project/src/project/resource-pack/animation-controller/animation-controller.d.ts
import { BaseObject } from 'bc-minecraft-bedrock-types';
import { MolangSet } from 'bc-minecraft-molang';
import { Using } from '../../../types/references';
/** */
export interface AnimationController extends BaseObject {
    /** */
    molang: MolangSet;
    /** */
    animations: Using;
    /** */
    particles: Using;
    /** */
    sounds: Using;
}
//# sourceMappingURL=animation-controller.d.ts.map